Bug#37462099 Bug#38001821 [MySql.EntityFrameworkCore 8.0.8 rename bug] [Contribution: fix: invalid cast exception]
This patch contains the fix of 2 bugs, one of which is a contribution and also that is partial fix for the other bug.

+++ EFCore/src/Migrations/Internal/MySQLHistoryRepository.cs
@@ -95,10 +95,10 @@ namespace MySql.EntityFrameworkCore.Migrations.Internal
       Dependencies.MigrationsLogger.AcquiringMigrationLock();
 
       var dbLock = CreateMigrationDatabaseLock();
-      int result;
+      Int64 result;
       try
       {
-        result = (int)(await CreateGetLockCommand().ExecuteScalarAsync(CreateRelationalCommandParameters(), cancellationToken)
+        result = (Int64)(await CreateGetLockCommand().ExecuteScalarAsync(CreateRelationalCommandParameters(), cancellationToken)
             .ConfigureAwait(false))!;
       }
       catch
